pyproject.toml
[project]
name = "media-monitoring-tool"
version = "0.1.0"
description = "Topic-focused media monitoring ingestion pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "feedparser>=6.0.11",
  "psycopg[binary]>=3.2.1",
  "python-dateutil>=2.9.0",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["media_monitoring", "media_monitoring.connectors", "media_monitoring.db"]
